+++ developer-extension/src/panel/components/tabs/diagnosticTab/data/custom.ts
@@ -1,26 +1,31 @@
-import type { Test } from '.'
+import type { Test } from '../../../../hooks/useTest'
 
 declare function asyncTestPassed(): void
 declare function asyncTestFailed(): void
 
 export const data: Test[] = [
   {
-    name: 'trusted events',
-    category: 'datadog',
-    exec() {
-      const request = new XMLHttpRequest()
-      // TODO pot to proxy or to datadog intake to go around CSP
-      request.open('POST', 'https://7fc280bc.datadoghq.com/api/v2/logs', true)
+    name: 'XMLHttpRequest',
+    subtests: [
+      {
+        name: 'generates trusted events',
+        category: 'datadog',
+        exec() {
+          const request = new XMLHttpRequest()
+          // TODO pot to proxy or to datadog intake to go around CSP
+          request.open('POST', 'https://7fc280bc.datadoghq.com/api/v2/logs', true)
 
-      request.addEventListener(
-        'loadend',
-        (event) => {
-          event.isTrusted ? asyncTestPassed() : asyncTestFailed()
+          request.addEventListener(
+            'loadend',
+            (event) => {
+              event.isTrusted ? asyncTestPassed() : asyncTestFailed()
+            },
+            { once: true }
+          )
+
+          request.send('')
         },
-        { once: true }
-      )
-
-      request.send('')
-    },
+      },
+    ],
   },
 ]
