+++ developer-extension/src/panel/components/tabs/diagnosticTab/data/index.ts
@@ -1,18 +1,9 @@
+import type { Test } from '../../../../hooks/useTest'
 import { data as custom } from './custom'
 import { data as es5 } from './es5'
 import { data as es6 } from './es6'
 import { data as es2016plus } from './es2016plus'
 import { data as esnext } from './esnext'
-
-export interface Test {
-  name: string
-  category?: string
-  spec?: string
-  mdn?: string
-  significance?: string
-  subtests?: Test[]
-  exec?: () => void
-}
 
 export const data: Test[] = [
   {
