+++ developer-extension/src/panel/hooks/useTest/sanitize.ts
@@ -1,0 +1,3 @@
+export function sanitize(s: string) {
+  return s.replace(/[^a-zA-Z0-9 ]/g, '').replace(/ /g, '-')
+}
