+++ developer-extension/src/panel/hooks/useTest/type.ts
@@ -1,0 +1,9 @@
+export interface Test {
+  name: string
+  category?: string
+  spec?: string
+  mdn?: string
+  significance?: string
+  subtests?: Test[]
+  exec?: () => void
+}
